+++ client_src/utils.h
@@ -17,7 +17,7 @@ int startOneDaemon( char * const * daemonArgs);
 
 // Set up the CCI connection
 // Returns a cci_status value, or a negated errno value (ie: -22 for EINVAL)
-int initIo( void *buffer, uint32_t len, uint32_t rank);
+int initIo( void *buffer, uint64_t len, uint32_t rank);
 
 // Shut down the CCI connection
 // Note: doesn't touch the daemon.  (Presumably, it will shut itself down.)
